Extract unit tests to ~260ms faster unit test class
